alterando

+++ src/main/java/com/gilmarcarlos/developer/gcursos/service/eventos/categorias/CategoriaEventoService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.gilmarcarlos.developer.gcursos.model.eventos.categorias.CategoriaEvento;
+import com.gilmarcarlos.developer.gcursos.model.eventos.presencial.exceptions.CategoriaException;
 import com.gilmarcarlos.developer.gcursos.repository.eventos.categorias.CategoriaEventoRepository;
 
 @Service
@@ -18,7 +19,18 @@ public class CategoriaEventoService {
 		return repository.save(categoria);
 	}
 		
-	public void deletar(Long id) {
+	public void deletar(Long id) throws CategoriaException {
+		
+		CategoriaEvento categoria = buscarPor(id);
+		
+		if(categoria == null) {
+			throw new CategoriaException("categoria não existe");
+		}
+		
+		if(!categoria.getEventos().isEmpty()) {
+			throw new CategoriaException("existem eventos cadastrados com essa categoria");
+		}	
+		
 		repository.deleteById(id);
 	}
